requirements.txt
# AI/ML Dependencies
beautifulsoup4
plotly
transformers
sentence-transformers
datasets==3.6.0
openai
anthropic
google-generativeai
gradio
gensim
modal
ollama
psutil
setuptools
speedtest-cli
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-community
feedparser
twilio
pydub
protobuf==3.20.2
wandb
ipython
bs4
PyPDF2
python-docx

# Core Python Dependencies
python-dotenv
requests
numpy
pandas
scipy
matplotlib
scikit-learn
chromadb
pyarrow
fpdf

# Development Dependencies
jupyterlab
ipywidgets
jupyter-dash

# PyTorch (CPU version - users can install GPU version separately if needed)
--extra-index-url https://download.pytorch.org/whl/cpu
torch